🏗️ Architect: Modularized optics calculations
Fix: Split monolithic `apts/optics/calculations.py` into specialized modules within a `calculations/` package.
- `atmospheric.py`: Handles airmass, extinction, and dispersion calculations.
- `photometry.py`: Handles sky/object flux and SNR calculations.
- `exposure.py`: Handles NPF rule and field rotation rate calculations.
- `resolution.py`: Handles Airy disk diameter calculations.
- `eyepiece.py`: Handles exit pupil and brightness calculations.

Benefit: Improves maintainability and adheres to the Single Responsibility Principle while preserving the public API through `__init__.py` re-exports. Internal models now import from specific submodules to reduce coupling.
